# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/dogs/{id}": {
      "patch": {
        "operationId": "MeController_editDog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DogDto"
                }
              }
            }
          }
        },
        "summary": "Update a dog",
        "tags": [
          "me"
        ]
      },
      "delete": {
        "operationId": "MeController_removeDog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "summary": "Soft-delete a dog",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "ModiModi API",
    "description": "Two-sided dog-walking marketplace. Errors always use the envelope { error: { code, message, requestId } }.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "ბობი"
          },
          "breed": {
            "type": "string",
            "example": "ლაბრადორი"
          },
          "birthDate": {
            "type": "string",
            "example": "2023-03-14",
            "description": "Date only. The app renders \"3 წლის\" from this."
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "sizeKg": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "breed",
          "birthDate"
        ]
      },
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
...
    }
  }
}